Decouple RL framework from TMNF game integration (10-phase refactor)
Restructures the repository so game-agnostic RL algorithms live in
framework/ and all TMNF-specific code is a self-contained plugin in
games/tmnf/.  Adding a second game only requires creating games/<name>/
implementing the defined interfaces.

Key changes:
- framework/: obs_spec, base_env, base_reward, policies, training, analytics
  — fully game-agnostic; zero imports from games/
- games/tmnf/: all TMNF code (env, reward, clients, track, steering, lidar,
  obs_spec, actions, simple_policy, analytics, tools)
- Root files (main.py, analytics.py, policies.py, etc.) become thin shims
  re-exporting from new locations so all existing import paths still work
- main.py slimmed to ~100 lines; all loop logic in framework/training.py
- grid_search.py updated to import train_rl from framework.training
- Policy classes parameterised with ObsSpec + head_names / discrete_actions
  so they are no longer hardcoded to 3-action TMNF control
- Abstract bases: BaseGameEnv(gymnasium.Env, ABC), RewardCalculatorBase(ABC)
- All 97 tests pass; framework/ contains zero game-specific source imports
