+++ games/__init__.py
@@ -1,0 +1,5 @@
+"""Game integration plugins.
+
+Each sub-package (e.g. games.tmnf) implements the framework's abstract
+interfaces for a specific game engine.
+"""
