+++ games/tmnf/__init__.py
@@ -1,0 +1,4 @@
+"""TMNF (Trackmania Nations Forever) game integration.
+
+Implements the framework's abstract interfaces using the TMInterface API.
+"""
