+++ games/tmnf/clients/__init__.py
@@ -1,0 +1,1 @@
+"""TMNF TMInterface client implementations."""
