+++ games/tmnf/tools/__init__.py
@@ -1,0 +1,1 @@
+"""TMNF offline tools (data preparation, debugging)."""
